+++ packages/styles/src/index.ts
@@ -8,11 +8,13 @@ export { button } from "./packages/atomic/button";
 export { checkbox } from "./packages/atomic/checkbox";
 export { chip } from "./packages/atomic/chip";
 export { icon } from "./packages/atomic/icon";
+export { iconButton } from "./packages/atomic/iconButton";
 export { link } from "./packages/atomic/link";
 export { skeleton } from "./packages/atomic/skeleton";
 export { status } from "./packages/atomic/status";
 export { text } from "./packages/atomic/text";
 export { title } from "./packages/atomic/title";
+export { toast } from "./packages/atomic/toast";
 export { toggle } from "./packages/atomic/toggle";
 
 // types
@@ -26,5 +28,7 @@ export type { TitleSprinkle } from "./packages/atomic/title";
 export type { ButtonVariants } from "./packages/atomic/button";
 export type { CheckboxVariants } from "./packages/atomic/checkbox";
 export type { ChipVariants } from "./packages/atomic/chip";
+export type { IconButtonVariants } from "./packages/atomic/iconButton";
 export type { StatusVariants } from "./packages/atomic/status";
+export type { ToastVariants } from "./packages/atomic/toast";
 export type { ToggleVariants } from "./packages/atomic/toggle";